requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
asttokens==3.0.1
asyncio==4.0.0
attrs==25.4.0
cffi==2.0.0
colorama==0.4.6
comm==0.2.3
comtypes==1.4.15
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
discord.py==2.6.4
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
keyboard==0.13.5
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mouse==0.7.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==26.0
pandas==3.0.1
parso==0.8.6
pillow==12.1.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
pure_eval==0.2.3
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pywin32==311
pywinauto==0.6.9
pyzmq==27.1.0
setuptools==70.2.0
six==1.17.0
sounddevice==0.5.3
soundfile==0.13.1
stack-data==0.6.3
sympy==1.14.0
torch==2.10.0+cu126
torchvision==0.25.0+cu126
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.6.0
yarl==1.22.0
